#1935 Moved a test class into a more suitable package
